+++ elixir/.formatter.exs
@@ -1,3 +1,3 @@
 [
-  inputs: ["mix.exs", "{lib,test,scripts}/**/*.{ex,exs}"]
+  inputs: ["mix.exs", "{lib,test,scripts,guard}/**/*.{ex,exs}"]
 ]